Module avmath::calculations

Calculations related to the ICAO Standard Atmosphere

As described in the ICAO document 7488: Manual of the ICAO Standard Atmosphere. Calculations requiring information about atmospheric layers are valid for geopotential altitudes from 5 km below to 80 km above mean sea level (geometric altitudes: 4.996 km below to 81.020 km above)

Functions

collision_frequency
Frequency of collisions between particles of dry air within an area

dynamic_viscosity
Dynamic viscosity between two neighboring layers of dry air moving at different speeds given a temperature

ideal_pressure
The ideal pressure of a gass given density, specific heat capacity, and ambient temperature

kinematic_viscosity
Kinematic viscosity given the dynamic viscosity and density

mean_free_path_number_density
Mean distance between collisions of dry air given a number density

mean_free_path_temp_pres
Mean free path given temperature and pressure

mean_particle_speed
Mean particle speed in dry air

number_density
Density of elements within a volume given pressure and temperature

relative_humidity
Calculates the relative humidity given the ambient pressure and partial pressure of water vapor

saturation_vapor_pressure_fast
Calculates the saturation pressure of water vapor at a given thermodynamic temperature using an approximation

saturation_vapor_pressure_wobus
Calculates the saturation pressure of water vapor at a given thermodynamic temperature

specific_weight
Specific weight given density and gravitational acceleration

speed_of_sound
The speed of sound in dry air given an ambient temperature

standard_density_dry_air
Density of dry air at a given temperature and pressure

standard_pressure
Computes the standard pressure for a given altitude

standard_temperature
Computes the standard temperature

standard_temperature_in_layer
Computes the standard temperature at a particular altitude from attributes of the layer

thermal_conductivity
Thermal conductivity of between two layers of dry air with a given temperature difference

virtual_temperature
Computes the virtual temperature given the relative humidity