Module avmath::constants

Constant values used as the basis for many calculations

Functions

Bs
Dry air constant for dynamic viscosity calculations (βs) in kg / m·s·K½

Kappa
Adiabatic index of dry air (κ)

Md
Molecular weight of dry air (Md or M)

Mv
Molecular weight of water vapor (Mv)

R
Ideal gas constant (R)

Rd
Specific heat capacity of dry air (R / Md)

Rd_over_standard_gravity_msl
Inverse lapse rate of dry air due to the effects of gravity (Rd / g₀)

Rv
Specific heat capacity of water vapor (Rv / Mv)

S
The Sutherland temperature (S)

Sigma
Effective collision diameter of an air molecule (σ)

SigmaSquared
The square of the effective collision diameter of an air molecule (σ²)

avogadros_number
Avogadro's number (NA)

dry_air_relative_humidity
Relative humidity of dry air

earth_radius
Radius of the earth (r)

speed_of_sound_msl
Speed of sound in standard atmosphere at mean sea level (a₀)

standard_density_msl
Standard density of dry air at mean sea level (ρ₀)

standard_geopotential_metre
Standard geopotential metre (m′)

standard_gravity_msl
Standard gravitational acceleration at mean sea level (g₀)

standard_gravity_msl_over_Rd
Lapse rate of dry air due to the effects of gravity (g₀ / Rd)

standard_pressure_msl
Standard pressure of dry air at mean sea level (P₀)

standard_temperature_msl
Standard temperature of dry air at mean sea level (T₀)

water_freezing_point_msl
Standard temperature of the freezing point of water at mean sea level (Ti)